# docs/openapi.json
{
  "components": {
    "parameters": {
      "ListAfter": {
        "description": "Return only entries with id strictly greater than this. Use the `next` field from the previous response to advance.",
        "in": "query",
        "name": "after",
        "required": false,
        "schema": {
          "minimum": 0,
          "type": "integer"
        }
      },
      "ListLimit": {
        "description": "Page size. Default 100, max 1000. 0 is rejected.",
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "default": 100,
          "maximum": 1000,
          "minimum": 1,
          "type": "integer"
        }
      },
      "SnapshotIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "minimum": 0,
          "type": "integer"
        }
      },
      "VmIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "minimum": 0,
          "type": "integer"
        }
      }
    },
    "schemas": {
      "CreateVmRequest": {
        "properties": {
          "cmdline": {
            "type": "string"
          },
          "kernel": {
            "type": "string"
          },
          "memory_mib": {
            "default": 128,
            "minimum": 1,
            "type": "integer"
          },
          "rootfs": {
            "type": "string"
          },
          "snapshot_dir": {
            "type": "string"
          },
          "vcpus": {
            "default": 1,
            "minimum": 1,
            "type": "integer"
          },
          "vsock_cid": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ExecRequest": {
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "cwd": {
            "type": "string"
          },
          "env": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "type": "array"
          },
          "program": {
            "type": "string"
          },
          "timeout_ms": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "program"
        ],
        "type": "object"
      },
      "ExecResponse": {
        "properties": {
          "duration_ms": {
            "minimum": 0,
            "type": "integer"
          },
          "exit_code": {
            "type": "integer"
          },
          "signal": {
            "type": "integer"
          },
          "stderr": {
            "type": "string"
          },
          "stdout": {
            "type": "string"
          }
        },
        "required": [
          "stdout",
          "stderr",
          "duration_ms"
        ],
        "type": "object"
      },
      "FileReadResponse": {
        "properties": {
          "content": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "content"
        ],
        "type": "object"
      },
      "FileWriteRequest": {
        "properties": {
          "content": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "mode": {
            "minimum": 0,
            "type": "integer"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "type": "object"
      },
      "FileWrittenResponse": {
        "properties": {
          "bytes": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "bytes"
        ],
        "type": "object"
      },
      "ForkResponse": {
        "properties": {
          "fork_count": {
            "minimum": 0,
            "type": "integer"
          },
          "fork_ms": {
            "minimum": 0,
            "type": "integer"
          },
          "fork_total_ms": {
            "minimum": 0,
            "type": "integer"
          },
          "vm": {
            "$ref": "#/components/schemas/VmHandleDto"
          }
        },
        "required": [
          "vm",
          "fork_ms",
          "fork_count",
          "fork_total_ms"
        ],
        "type": "object"
      },
      "SnapshotDto": {
        "properties": {
          "dir": {
            "type": "string"
          },
          "display": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display"
        ],
        "type": "object"
      },
      "SnapshotListEntry": {
        "properties": {
          "display": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "kernel_cmdline": {
            "type": "string"
          },
          "memory_bytes": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "vcpu_count": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "display"
        ],
        "type": "object"
      },
      "SnapshotListResponse": {
        "properties": {
          "next": {
            "description": "Cursor for the next page. Pass as `?after=<this>`. Absent when the response holds the tail of the result set.",
            "minimum": 0,
            "type": "integer"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/SnapshotListEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "snapshots"
        ],
        "type": "object"
      },
      "SnapshotRequest": {
        "properties": {
          "to_dir": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VmHandleDto": {
        "properties": {
          "display": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/VmStateDto"
          }
        },
        "required": [
          "id",
          "display",
          "state"
        ],
        "type": "object"
      },
...
      "VmListResponse": {
        "properties": {
          "next": {
            "description": "Cursor for the next page. Pass as `?after=<this>`. Absent when the response holds the tail of the result set.",
            "minimum": 0,
            "type": "integer"
          },
          "vms": {
            "items": {
              "$ref": "#/components/schemas/VmListEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "vms"
        ],
        "type": "object"
      },
...
      "VmStateResponse": {
        "properties": {
          "display": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/VmStateDto"
          }
        },
        "required": [
          "id",
          "display",
          "state"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "rust-nano-vm control-plane API",
    "version": "0.0.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/snapshots": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ListLimit"
          },
          {
            "$ref": "#/components/parameters/ListAfter"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotListResponse"
                }
              }
            },
            "description": "Snapshot list"
          },
          "400": {
            "description": "limit out of range (0 or > 1000)"
          }
        },
        "summary": "List snapshots"
      }
    },
    "/v1/snapshots/{id}": {
      "delete": {
        "parameters": [
          {
            "$ref": "#/components/parameters/SnapshotIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "summary": "Delete snapshot"
      }
    },
    "/v1/snapshots/{id}/fork": {
      "post": {
        "description": "MAP_PRIVATE CoW fork of the snapshot. Subject to per-token token-bucket quota; throttled callers get 429 with Retry-After.",
        "parameters": [
          {
            "$ref": "#/components/parameters/SnapshotIdPath"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForkResponse"
                }
              }
            },
            "description": "Forked child VM plus per-fork latency and caller usage totals"
          },
          "429": {
            "description": "Rate-limited by per-token fork quota"
          }
        },
        "summary": "Fork a child VM from a snapshot (metered)"
      }
    },
    "/v1/snapshots/{id}/restore": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/SnapshotIdPath"
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VmHandleDto"
                }
              }
            },
            "description": "Created VM from snapshot"
          }
        },
        "summary": "Restore snapshot into VM"
      }
    },
...
    "/v1/vms": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ListLimit"
          },
          {
            "$ref": "#/components/parameters/ListAfter"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VmListResponse"
                }
              }
            },
            "description": "VM list"
          },
          "400": {
            "description": "limit out of range (0 or > 1000)"
          }
        },
        "summary": "List VMs"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVmRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VmHandleDto"
                }
              }
            },
            "description": "Created VM"
          }
        },
        "summary": "Create VM"
      }
    },
    "/v1/vms/{id}": {
      "delete": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Destroyed"
          }
        },
        "summary": "Destroy VM"
      },
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VmStateResponse"
                }
              }
            },
            "description": "VM state"
          }
        },
        "summary": "Get VM state"
      }
    },
    "/v1/vms/{id}/exec": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResponse"
                }
              }
            },
            "description": "Command result"
          }
        },
        "summary": "Execute a command in the guest"
      }
    },
    "/v1/vms/{id}/files": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReadResponse"
                }
              }
            },
            "description": "File content"
          }
        },
        "summary": "Read a file from the guest filesystem"
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileWrittenResponse"
                }
              }
            },
            "description": "Write result"
          }
        },
        "summary": "Write a file into the guest filesystem"
      }
    },
    "/v1/vms/{id}/snapshot": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDto"
                }
              }
            },
            "description": "Created snapshot"
          }
        },
        "summary": "Create snapshot"
      }
    },
    "/v1/vms/{id}/start": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Started"
          }
        },
        "summary": "Start VM"
      }
    },
    "/v1/vms/{id}/stop": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/VmIdPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Stopped"
          }
        },
        "summary": "Stop VM"
      }
    }
  }
}